file restructuring; no more airflow
